+++ lib/src/typedefs/on_tab_selection.dart
@@ -1,4 +1,4 @@
 import '../tab_data.dart';
 
 /// Event that will be triggered when the tab selection is changed.
-typedef OnTabSelection = void Function(TabData? tabData);
+typedef OnTabSelection = void Function(int tabIndex, TabData? tabData);